Refactor app into index, app and home pages
