+++ pages/__init__.py
@@ -1,0 +1,2 @@
+from . import home
+from . import champion
